+++ .babelrc
@@ -1,0 +1,8 @@
+
+{
+    "env": {
+      "test": {
+        "plugins": ["@babel/plugin-transform-modules-commonjs"]
+      }
+    }
+  }
